docker-compose.yml: remove mount for node_modules/.cache
Because if this mount is present, npm ci yields an error that
it cannot delete the node_modules/.cache directory.
Also remove frontend/node_modules/.gitkeep and print/node_modules/.gitignore,
because they are deleted by npm ci of frontend and print.
npm v7 keeps a package-lock.json in node_modules, which tracks
the content of the node_modules directory, and deletes
everything which should not be there according to it and
the package-lock.json in node_modules/..
This was added in 6d67db4.

Issue: #1390